+++ cmd/zetae2e/local/local.go
@@ -207,10 +207,9 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	// run setup steps that do not require tss
 	if !skipSetup {
 		noError(deployerRunner.FundEmissionsPool())
-		logger.Print("running gov proposals")
 	}
 
-	noError(deployerRunner.CreateGovProposal(runner.StartOfE2E))
+	noError(deployerRunner.CreateGovProposals(runner.StartOfE2E))
 
 	// wait for keygen to be completed
 	// if setup is skipped, we assume that the keygen is already completed
@@ -493,7 +492,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 		logger.Print("❌ e2e tests failed after %s", time.Since(testStartTime).String())
 		os.Exit(1)
 	}
-	noError(deployerRunner.CreateGovProposal(runner.EndOfE2E))
+	noError(deployerRunner.CreateGovProposals(runner.EndOfE2E))
 
 	logger.Print("✅ e2e tests completed in %s", time.Since(testStartTime).String())